backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1

# Document processing - Updated versions with security fixes
PyMuPDF==1.23.26
pytesseract==0.3.10
Pillow==10.2.0
nltk==3.8.1

# Vector database - Compatible versions
qdrant-client>=1.9.0,<2.0.0
langchain-qdrant==0.1.0

# Embeddings and LangChain - Use newer compatible versions
langchain==0.1.20
langchain-community==0.0.38
langchain-google-genai==1.0.1
langchain-openai==0.1.0
langchain-text-splitters==0.0.1
langsmith>=0.1.0,<0.2.0  # Use newer compatible langsmith version
google-generativeai>=0.4.0

# LLM - Updated version
groq==0.4.2

# Core dependencies - Updated versions
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
aiofiles==23.2.1

# Additional useful packages for production
httpx==0.27.0  # Better HTTP client for async operations
tenacity==8.2.3  # Retry logic for external API calls

# Security and validation
cryptography==42.0.2  # For secure operations
validators==0.22.0  # Input validation utilities

# Development and debugging tools (removed some to reduce conflicts)
pytest==7.4.4
pytest-asyncio==0.23.4

# Production utilities
gunicorn==21.2.0  # Production WSGI server
uvloop==0.19.0  # High-performance event loop
httptools==0.6.1  # Fast HTTP parsing